Creat a build that always starts with the light on
I chose level 80 for the default as with my TS11 this level consistently
works on many chargers with no battery inserted. You might be able to
tune it a bit higher, but going higher than the charging circuit can
deliver will cause LVP. LVP could theoretically be disabled but I'm
unsure about doing this in case it purs stress on the charging circuit.

If you want to adjust the starting brightness then 10C to save
brightness, all the normal anduril config still works and I also made
the light start in advanced UI by default to make saving with 10C
simpler, as well as configuring aux.